+++ samples/main.cpp
@@ -1,0 +1,8 @@
+#include <iostream>
+#include "TBitField.h"
+
+int main()
+{
+	TBitField K, S, M;
+	 
+}
