+++ src/CMakeLists.txt
@@ -1,6 +1,6 @@
-set(target ${BIT_LIBRARY})
+set(target ${MP2_LIBRARY})
 
-file(GLOB hdrs "*.h*" "${BIT_INCLUDE}/*.h*")
+file(GLOB hdrs "*.h*" "${MP2_INCLUDE}/*.h*")
 file(GLOB srcs "*.cpp")
 
 add_library(${target} STATIC ${srcs} ${hdrs})
