+++ test/CMakeLists.txt
@@ -1,4 +1,4 @@
-set(target ${BIT_TESTS})
+set(target ${MP2_TESTS})
 
 file(GLOB hdrs "*.h*")
 file(GLOB srcs "*.cpp")
@@ -6,4 +6,4 @@ file(GLOB srcs "*.cpp")
 include_directories("${CMAKE_CURRENT_SOURCE_DIR}/../3rdparty")
 
 add_executable(${target} ${srcs} ${hdrs})
-target_link_libraries(${target} gtest ${BIT_LIBRARY})
+target_link_libraries(${target} gtest ${MP2_LIBRARY})
